worker: caches with request again
There is a weird behaviour with the `Origin` header.
If present it seems like the resource (response) gets
cached with its own key.
Chrome requests the js and wasm files with an origin
header bypassing the cache if the resource key is just the url.
Requests with an origin header can also not be purged from the
ui by url.